Improve scoresd.py's handling of starting and stopping a bit
This is all aimed at the issue that it takes 3-4 hours on CAO to do a
rebuild of player pages when the daemon restarts, even though this is
almost completely unnecessary.

This includes three main changes:
 * do not rebuild the entire player pages directory on every daemon
   restart. Summary pages are still rebuilt, and player pages that are
   dirty because of any incremental updates are as well.
 * add some options to let someone force a full rebuild if needed.
 * handle shutdown when there are dirty player pages better, so that
   there is less need to manually rebuild all pages. Specifically, flush
   any dirty pages on SIGTERM -- before, these were just dropped.

There's some light refactoring of the page update code as part of this,
though it's overall still pretty wild.