Tweak timing for some heavy pages
Also add a caveat about inactive accounts to those pages.

As noted in the comment, these values are kind of weird to interact with
because they depend on how long a tick takes. At the time of this
commit, on CAO it is about 7 minutes of activity + 1 of sleep -- but,
the tick count for these particular pages is incremented pretty much any
time a game ends, so they also depend on how much players are doing. In
my testing, during non-tournament times, these two pages get incremented
about 20-30 times a tick. So, this commit *should* entail them getting
updated at most once an hour, in contrast to ~ every 15 minutes?

As a side note, these calculations suggest that 30 ticks as a default
value for player pages is way off -- unless a player wins this means
once every 3 hrs, as opposed to once every 30 minutes as the comments
there indicate.