Attempt to normalize ticks based on time
This code was originally designed when scoring was not io bound, and
most time was spent sleeping, I think. This is not what things are like
in the current setup; this change will hopefully ensure the original
idea that every dirty page gets updated at least once every 30 minutes
or so.