Add a new flag `dormant`, for dead servers
You can accomplish something similar with `local`, but you then need to
in many cases carefully edit the logfile names in sources so that they
match up to what's in the database; in addition `local` has some
behavior that is intended for actively updated files. Setting `dormant
to `True` just very straightforwardly disables the fetch step and any
updating to the logfile, while using the cached logfile in its current
location when doing any rereading. N.b. this won't prevent scoresd from
reading log entries from the cached logfile, if the database file
location is behind the end of the file.

Flag is applied to CSZO and RHF.