Fix handling of sparse data in per-day graph generation
If a day is completely empty it is not returned in the query used to
generate these graphs. Previously, the data point was also dropped from
the graph, leading to uneven x axes for sparse data sets. This doesn't
come up so much on CAO scoring but may have been having subtle effects;
it definitely matters for graphing other data sets, for example activity
for one logfile on a lightly-used server. This change just fills in the
missing dates, and also doesn't crash if a graph is completely empty.
(Is there a way to fill in dates in sql?)