Fix a bug in how logfile offsets are stored in the db
The original implementation had much more apparent behavior:
XLogFile.offset was always the offset of the last read line, and this
was stored in the db. At some point, things got more complicated, and
XLogfile.offset was actually the next read point -- but now this got
stored in the db, and the code to load offsets from the db didn't get
correspondingly changed. Consequently, on daemon restart, if a logfile
had any new loglines, the first one would get dropped. (If it didn't
there were no consequences.) This also caused loglines that were
incomplete on the last wget to be dropped.

In principle, the db could either store the next read point or the last
read line, but I decided the latter was safer. (Unfortunately this
entails one more risky daemon restart on CAO.)

This is the bug that caused #2, just by extremely unlucky timing - the
logline was incomplete by character at the time of the wget.