Add a command to rebuild specific player pages
For now, this is just an extra startup step on a full scoresd load.

This commit also fixes a crash in a corner case where a logfile
consisted of a single line; I'm not sure what the original intent of the
`seek` call here was, but it definitely doesn't work now if offset is 0.