Convert some INSERTs to be multiple-valued
In general, a problem with logfile reading is that there are many, many
individual INSERT statements per logline (interspersed with other
queries in some cases). This commit begins to chip away at this for one
table, by saving up the rows to be inserted/updated until just before a
commit for two tables storing the per-day stats. I'm not sure this
commit will make that much difference on its own, but these are two of
the easiest tables to convert to a multiple-value INSERT, so it's a
useful starting point.