Only update all_recent_games once per commit, using bulk INSERT
This was updated ncrementally for every single game processed, even
though it maxes out at 100 games and even a single commit processes
3000. This change only does this insert when a db commit is going to
happen. I've also changed things so that as long as the table is full
(which for practical purposes is always true) during a bulk import it
will use TRUNCATE to empty the table.