Do streak-related db updates once per commit
Instead of doing the streak processing and db insertion at each game,
this now does the processing in python only and the insertion at the end
of each commit cycle. Because this code is already complicated enough, I
didn't try to do any kind of bulk INSERT with the streak data -- but
nonetheless, this gives a huge speedup, halving (or more) the time taken
to import my test data. I think most of the cost was in the processing,
which was very db-intensive before, and intermixed with updates; but
most of the interim per-game processing can be done in python. This code
is as tested as I can get it but it will need some scrutiny on the full
data set, as it is definitely more complex than what it is replacing.