Fix slow sorting with many logfiles
This turned out to be extremely inefficient, taking something like 1.5x
the actual db loading stuff with the full logfile set in my local
testing. I'm guessing this was envisioned with only a few logfiles,
where the difference between doing a full sort and doing an insort is
negligible.

A few other tweaks in this commit.