feat: more usable db version info
This commit caches coarse version info in the database in a much more
usable format than just the usual `v` etc fields. This inspired by how
some version info is handled in sequell, but I have structured in (what
seems to me) to be a bit more of a straightforward way with respect to
ideas that I have for the frontend. A lot of this is on spec, but I've
added some stuff to overview.mako that uses it.

Obviously, this commit changes the db schema and so isn't backwards
compatible. However, the table is reconstructed when pulling new data,
so it is safe to manually create the table according to the schema
without rebuilding the database. Just run the relevant CREATE TABLE
lines and add the version indices that come as part of this commit.