fix: optimize some of the worst page rendering
For each the total score and all-players pages, for each row returned in
the main query, there was a followup subquery to get that player's
best/first/last games. Needless to say, this did not scale. The main
change in this commit is to refactor this so that all this info is
collected in a single (inner join) query. This gets 1-2 orders of
magnitude speedup. The most unrestricted queries for this are still not
fast, but they are so much better -- e.g. the worst case scenario on a
very fast local render was over a minute for all this querying combined
(much worse when running on CAO), and this gets it down to a few
seconds. There's a bunch of subsidiary changes as well:

* I split the "best players" concept into one that only looks at the
  past 6 months, and one that looks at all players with total score >
  500. The 6mo version is correspondingly much faster again (about
  another order of magnitude from the previous wishy-washy restricted
  query.)
* The worst pages now run actually once per day. Because of this
  combined with the speedups, I've relaxed some of the restrictions (and
  all-players is again really all-players). We'll see how this goes on
  CAO.
* fixed a bug with manual page rendering and custom thresholds.
  added some explicit slow render logging (this info was already there,
  but this will make it a bit easier to find).