fix: correctly sort player_best_games in bulk inload
This has been broken for four years, since 818bd17c5d46d. Its direct
impact is somewhat minimal, but if key games are missing, it can now
result in players not showing up in the overall lists. (It looks like
this was previously only true for the winners page.)

A complete fix for this would need a full db reload. For now I have done
some surgery on CAO to make sure that all top winning games are present
in this table (where before, there were quite a few missing).