feat: basic setup for descent scoring
Sources (just one right now) and setup for descent mode scoring.

Aside from that, this is mostly cosmetic changes for when the scoring
repo is used for dungeon descent, rather than regular mode. (Right now,
it's not advisable to mix modes in the same running instance of a
scoring database -- this just uses a fresh database, does a minor batch
of retheming, etc)