fix: tweak timeout settings for log downloading
Previously, this did 60s with the default number of tries, which was 20,
and therefore took 20 minutes per logfile; if an entire server goes down
this multiplies out by the number of logfiles on that server. This code
should probably be smarter in some way (e.g. dynamically react to
failures) but for now I'm going to shorten the timeout and not retry --
it'll still get a retry on the next loop.