fix: mark cwz as dormant
This is hopefully temporary, but given the way that logfile downloading
works the fact that cwz logfiles time out right now imposes a huge delay
in updating scores for other servers. (Before the prior commit, it was
60s x 20 retries x 19 logfiles, so 6 hours per import loop...)

CWZ logfiles at this point have been inaccessible since 2023-08-07.
To fix this, I need working logfile urls. Presumably a similar thing
applies to sequell (which is likely smarter about timeouts, but may
still be getting some lag from this).